backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
sqlalchemy==2.0.49
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.0
pydantic[email]==2.11.5
